Move AttributeErrorCode to attribute module
