+++ saleor/attribute/error_codes.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class AttributeErrorCode(Enum):
+    ALREADY_EXISTS = "already_exists"
+    GRAPHQL_ERROR = "graphql_error"
+    INVALID = "invalid"
+    NOT_FOUND = "not_found"
+    REQUIRED = "required"
+    UNIQUE = "unique"
